+++ project/src/components/offers-list/offers-list.jsx
@@ -1,0 +1,10 @@
+import React from 'react';
+import Card from '../card/card';
+
+function OffersList(props) {
+  const { offers } = props;
+
+  return offers.map((offer) => <Card {...props} key={offer.id} offer={offer} />);
+}
+
+export default OffersList;
